:hammer: Move Format types into new assembly
