+++ yarhl.Media/Properties/AssemblyInfo.cs
@@ -1,0 +1,51 @@
+﻿//
+// AssemblyInfo.cs
+//
+// Author:
+//       Benito Palacios Sánchez <benito356@gmail.com>
+//
+// Copyright (c) 2017 Benito Palacios Sánchez
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+using System;
+using System.Reflection;
+using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
+using Mono.Addins;
+
+[assembly: AssemblyTitle("yarhl.Media")]
+[assembly: AssemblyDescription("Media formats for Yarhl")]
+[assembly: AssemblyCompany("yarhl development team")]
+[assembly: AssemblyProduct("yarhl")]
+[assembly: AssemblyCopyright("Copyright (c) 2017 Benito Palacios (aka pleonex)")]
+[assembly: AssemblyVersion("1.0.0.*")]
+[assembly: AssemblyFileVersion("1.0.0")]
+[assembly: CLSCompliant(true)]
+[assembly: ComVisible(true)]
+[assembly: InternalsVisibleTo("yarhl.UnitTests")]
+
+#if DEBUG
+[assembly: AssemblyConfiguration("Debug")]
+#elif RELEASE
+[assembly: AssemblyConfiguration("Release")]
+#endif
+
+// Mono.Addins
+[assembly: Addin("yarhl.Media", "1.0")]
+[assembly: AddinDependency("yarhl", "1.0")]
