+++ yarhl.Media/Text/Po.cs
@@ -23,13 +23,14 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.FileFormat.Common
+namespace Yarhl.Media
 {
     using System;
     using System.Collections.Generic;
     using System.Collections.ObjectModel;
     using System.Linq;
     using Mono.Addins;
+    using Yarhl.FileFormat;
 
     /// <summary>
     /// Portable Object format for translations.
