+++ yarhl.Media/Text/Po2Binary.cs
@@ -23,12 +23,12 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.FileFormat.Common
+namespace Yarhl.Media
 {
     using System;
-    using System.Linq;
     using System.Text;
     using Mono.Addins;
+    using Yarhl.FileFormat;
     using Yarhl.IO;
 
     /// <summary>
