+++ yarhl.Media/Text/PoEntry.cs
@@ -23,7 +23,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.FileFormat.Common
+namespace Yarhl.Media
 {
     /// <summary>
     /// Entry in PO formats. Represents a translation unit.
