+++ yarhl.UnitTests/FileFormat/BinaryFormatTests.cs
@@ -23,11 +23,11 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.FileFormat.Common
+namespace Yarhl.UnitTests.FileFormat
 {
     using System.IO;
     using NUnit.Framework;
-    using Yarhl.FileFormat.Common;
+    using Yarhl.FileFormat;
     using Yarhl.IO;
 
     [TestFixture]
