+++ yarhl.UnitTests/Media/Po2BinaryTests.cs
@@ -23,12 +23,13 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.FileFormat.Common
+namespace Yarhl.UnitTests.Media
 {
     using System;
     using NUnit.Framework;
-    using Yarhl.FileFormat.Common;
+    using Yarhl.FileFormat;
     using Yarhl.IO;
+    using Yarhl.Media;
 
     [TestFixture]
     public class Po2BinaryTests
