+++ yarhl.UnitTests/Media/PoEntryTests.cs
@@ -23,10 +23,10 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.UnitTests.FileFormat.Common
+namespace Yarhl.UnitTests.Media
 {
     using NUnit.Framework;
-    using Yarhl.FileFormat.Common;
+    using Yarhl.Media;
 
     [TestFixture]
     public class PoEntryTests
