+++ yarhl.sln
@@ -5,6 +5,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "yarhl", "yarhl\yarhl.csproj
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "yarhl.UnitTests", "yarhl.UnitTests\yarhl.UnitTests.csproj", "{3525724C-57EC-4C84-ABFC-95780009A304}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "yarhl.Media", "yarhl.Media\yarhl.Media.csproj", "{D518538F-7B50-41A0-839C-C1D6A287DFA3}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -19,6 +21,10 @@ Global
 		{47373F3F-6A6A-4B2A-B06E-971D9FBDB545}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{47373F3F-6A6A-4B2A-B06E-971D9FBDB545}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{47373F3F-6A6A-4B2A-B06E-971D9FBDB545}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D518538F-7B50-41A0-839C-C1D6A287DFA3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D518538F-7B50-41A0-839C-C1D6A287DFA3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D518538F-7B50-41A0-839C-C1D6A287DFA3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D518538F-7B50-41A0-839C-C1D6A287DFA3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		version = 1.0
