+++ yarhl/FileSystem/Node.cs
@@ -22,7 +22,6 @@ namespace Yarhl.FileSystem
 {
     using System;
     using FileFormat;
-    using FileFormat.Common;
 
     /// <summary>
     /// Node in the FileSystem with an associated format.
