+++ yarhl/FileSystem/NodeContainerFormat.cs
@@ -23,10 +23,10 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-namespace Yarhl.FileFormat.Common
+namespace Yarhl.FileSystem
 {
     using System;
-    using FileSystem;
+    using FileFormat;
     using Mono.Addins;
 
     /// <summary>
